project/apps/accounts/api/v1/serializers.py
from django.contrib.auth import get_user_model

from rest_framework import serializers


UserModel = get_user_model()


class UserCardSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserModel
        fields = ('pk', 'username', 'first_name', 'last_name')


class UserDetailSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserModel
        fields = ('pk', 'username', 'email', 'first_name', 'last_name', 'friends')
